% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{abort_dlw_error}
\alias{abort_dlw_error}
\title{return information of error}
\usage{
abort_dlw_error(info)
}
\arguments{
\item{info}{csv file with at least one variable with "error" in its name.}
}
\value{
error
}
\description{
return information of error
}
\keyword{internal}
