% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_server_catalog.R
\name{add_gmd_vars}
\alias{add_gmd_vars}
\title{Add GMD vars to GMD server catalog}
\usage{
add_gmd_vars(ctl)
}
\arguments{
\item{ctl}{data.table of server catalog raw}
}
\value{
data.table with more variables
}
\description{
Add GMD vars to GMD server catalog
}
\keyword{internal}
