% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{data_calls}
\alias{data_calls}
\title{get data calls when more than one option.}
\usage{
data_calls(ctl, country_code)
}
\arguments{
\item{ctl}{data.table from [dlw_server_inventory]}

\item{country_code}{ISO3}
}
\value{
list of possible calls
}
\description{
get data calls when more than one option.
}
\keyword{internal}
