% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_country_catalog.R
\name{dlw_country_catalog}
\alias{dlw_country_catalog}
\title{Get country catalog}
\usage{
dlw_country_catalog(
  country_code,
  dlw_url = NULL,
  api_version = getOption("dlw.default_api_version"),
  force = FALSE,
  internal = FALSE,
  verbose = getOption("dlw.verbose"),
  store_request = FALSE
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{dlw_url}{character: in case we have more than one url.}

\item{api_version}{character: API version}

\item{force}{logical: If TRUE, it will query the API regardless of whether
there is version available in .dlwenv}

\item{internal}{logical: for internal use ONLY.}

\item{verbose}{logical: whether to display info}

\item{store_request}{logical: store request in .dlwenv as `last_req`}
}
\value{
data.table with country catalog
}
\description{
Get country catalog
}
\examples{
\dontrun{
dlw_country_catalog("COL")
}
}
