% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd.R
\name{dlw_get_gmd}
\alias{dlw_get_gmd}
\title{Get GMD data.}
\usage{
dlw_get_gmd(
  country_code,
  year = NULL,
  module = NULL,
  survey = NULL,
  filename = NULL,
  vermast = NULL,
  veralt = NULL,
  latest_version = TRUE,
  latest_year = FALSE,
  verbose = getOption("dlw.verbose"),
  ...
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{year}{numeric: four digit year}

\item{module}{character: module of GMD collection (e.g., ALL, GPWG, L)}

\item{survey}{character: survey acronyn}

\item{filename}{character: File name}

\item{vermast}{character: Version of the master data in the form "vXX" where
X is a number of two digits like "01" or "02".}

\item{veralt}{character: Version of the alternative  data in the form "vXX"
where X is a number of two digits like "01" or "02".}

\item{latest_version}{logical: If TRUE and  `vermast` and `veralt` are NULL,
it will use the most recent version of the data for a particular year.}

\item{latest_year}{logical: If `TRUE` and  `year` is NULL, it retrieves the
most recent year. Otherwise, it will return the calls for all the years
available. This is the default.}

\item{verbose}{logical: whether to display info}

\item{...}{
  Arguments passed on to \code{\link[=dlw_get_data]{dlw_get_data}}
  \describe{
    \item{\code{local_dir}}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}
    \item{\code{local}}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}
    \item{\code{local_overwrite}}{logical. Whether to overwrite any saved data. Default
is FALSE}
  }}
}
\value{
If the call is unique, it will return the data. If not, it will
  return the posibilities for the user to choose.
}
\description{
This is a wrapper of dlw_get_data. It just build the calls to dlw_get_data,
for ease of use.
}
\examples{
\dontrun{
  dlw_get_gmd("PRY", 2010, "GPWG") # latest version of 2010
  dlw_get_gmd("PRY", module = "GPWG") # latest year and latest version
}
}
