% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd.R
\name{dlw_get_gmd_support}
\alias{dlw_get_gmd_support}
\title{get Support files for GMD}
\usage{
dlw_get_gmd_support(
  module = c("CPIICP", "CPI"),
  vermast = NULL,
  verbose = getOption("dlw.verbose")
)
}
\arguments{
\item{module}{character: As of now, either CPIICP (the default) or CPI.}

\item{vermast}{character: Version of the master data in the form "vXX" where
X is a number of two digits like "01" or "02".}

\item{verbose}{logical: whether to display info}
}
\value{
data.table
}
\description{
get Support files for GMD
}
\examples{
\dontrun{
dlw_get_gmd_support()
dlw_get_gmd_support(vermast = "v10")
}
}
\seealso{
Other GMD utilities: 
\code{\link{dlw_get_gmd}()},
\code{\link{eval_gmd_call}()},
\code{\link{gmd_calls}()}
}
\concept{GMD utilities}
