% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_pin_info.R
\name{dlw_get_pins_versions}
\alias{dlw_get_pins_versions}
\title{Retrieve and Save Version Information for GMD Pins}
\usage{
dlw_get_pins_versions(
  local_dir = fs::path(Sys.getenv("PIP_ROOT_DIR"),
    "PIP_ingestion_pipeline_v2/dlw_repository/dlw_data"),
  local_info = fs::path(Sys.getenv("PIP_ROOT_DIR"),
    "PIP_ingestion_pipeline_v2/dlw_repository/dlw_inventory")
)
}
\arguments{
\item{local_dir}{character: Local directory where the pins are saved.}

\item{local_info}{character: Local directory where the pins information is saved.}
}
\value{
Invisibly returns `NULL`. The function is called for its side effects:
saving the pin metadata to the inventory board.
}
\description{
This function retrieves metadata (including versions, creation dates, and hashes)
for all pins stored in the `dlw_data` board, processes and orders the metadata,
and writes the resulting summary to the `dlw_inventory` board as a versioned pin.
}
\examples{
\dontrun{
dlw_get_pins_versions()
}
}
