% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlwenv.R
\name{dlw_list_env}
\alias{dlw_list_env}
\title{Get the list of objects in a specified environment}
\usage{
dlw_list_env(env = .dlwenv)
}
\arguments{
\item{env}{A Environment name, default to `.dlwenv`}
}
\value{
List of objects
}
\description{
Get the list of objects in a specified environment
}
\examples{
dlw_list_env()
}
