% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_pin_info.R
\name{dlw_read_pins_versions}
\alias{dlw_read_pins_versions}
\title{Retrieve the DLW GMD Pins Version Info}
\usage{
dlw_read_pins_versions(
  local_info = fs::path(Sys.getenv("PIP_ROOT_DIR"),
    "PIP_ingestion_pipeline_v2/dlw_repository/dlw_inventory")
)
}
\arguments{
\item{local_info}{character: Local directory where the pins information is saved.}
}
\value{
A `data.table` containing pin metadata including pin name, version,
creation date, and hash.
}
\description{
This function reads the pin containing metadata about DLW GMD pins
(written by `dlw_get_pins_versions()`) from the `dlw_inventory` board.
}
\examples{
\dontrun{
pin_info <- dlw_read_pins_versions()
head(pin_info)
}
}
