% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_server_catalog.R
\name{dlw_server_inventory}
\alias{dlw_server_inventory}
\title{filter server catolog to show inventory by country and other variables}
\usage{
dlw_server_inventory(country, server = NULL, ...)
}
\arguments{
\item{country}{character. same as `country_code` in [dlw_get_data] but with
the purpose of easy programming. Not meant to be used by final user.}

\item{server}{character: in case we have more than one server. default is GMD}

\item{...}{additional filtering arguments (e.g., year, module, survey, fileName)}
}
\value{
filter server catalog from [dlw_server_catalog]
}
\description{
filter server catolog to show inventory by country and other variables
}
\examples{
\dontrun{
dlw_server_inventory("COL")
dlw_server_inventory("COL", year = 2010)
}
}
\seealso{
Other catalogs: 
\code{\link{dlw_country_catalog}()},
\code{\link{dlw_server_catalog}()}
}
\concept{catalogs}
