% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_utils.R
\name{dlw_test_token}
\alias{dlw_test_token}
\title{confirm wither the token is working correctly}
\usage{
dlw_test_token(
  dlw_url = NULL,
  api_version = getOption("dlw.default_api_version"),
  verbose = getOption("dlw.verbose")
)
}
\arguments{
\item{dlw_url}{character: in case we have more than one url.}

\item{api_version}{character: API version}

\item{verbose}{logical: whether to display info}
}
\value{
TRUE if all goes well
}
\description{
confirm wither the token is working correctly
}
\examples{
\dontrun{
dlw_test_token()
}
}
