% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{.dlwenv}
\alias{.dlwenv}
\title{DLW environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.dlwenv
}
\description{
This environment stores all active  files used internally by the package.
}
\keyword{internal}
