% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd.R
\name{eval_gmd_call}
\alias{eval_gmd_call}
\title{evaluate GMD call}
\usage{
eval_gmd_call(calls, verbose = getOption("dlw.verbose"))
}
\arguments{
\item{calls}{calls from [gmd_calls]}

\item{verbose}{logical: whether to display info}
}
\value{
evaluation
}
\description{
evaluate GMD call
}
\seealso{
Other GMD utilities: 
\code{\link{dlw_get_gmd}()},
\code{\link{dlw_get_gmd_support}()},
\code{\link{gmd_calls}()}
}
\concept{GMD utilities}
\keyword{internal}
