% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlwenv.R
\name{get_dlwenv}
\alias{get_dlwenv}
\title{Get function: Returns the entire .dlwenv environment}
\usage{
get_dlwenv()
}
\value{
The .dlwenv environment
}
\description{
Get function: Returns the entire .dlwenv environment
}
\examples{
env <- get_dlwenv()
}
