% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlwenv.R
\name{get_from_dlwenv}
\alias{get_from_dlwenv}
\title{Get for a specific key from .dlwenv}
\usage{
get_from_dlwenv(key, verbose = getOption("dlw.verbose"))
}
\arguments{
\item{key}{A character string representing the key}
}
\value{
The value associated with the key in .dlwenv
}
\description{
Get for a specific key from .dlwenv
}
\examples{
set_in_dlwenv("example_key", 42)
get_from_dlwenv("example_key") # returns 42
}
