% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{perform request and get raw data}
\usage{
get_raw_data(req)
}
\arguments{
\item{req}{request from [build_request]}
}
\value{
raw data from [resp_body_raw]
}
\description{
perform request and get raw data
}
\keyword{internal}
