% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{handle_resp}
\alias{handle_resp}
\title{perform httr2::req_perform and handle errors properly}
\usage{
handle_resp(req)
}
\arguments{
\item{req}{A httr2 request object.}
}
\value{
an HTTP response.
}
\description{
perform httr2::req_perform and handle errors properly
}
