% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{load_from_local}
\alias{load_from_local}
\title{Load from local drive}
\usage{
load_from_local(local_dir, filename)
}
\arguments{
\item{local_dir}{local directory}

\item{filename}{File name. "dta" format for now}
}
\value{
data in `fs::path(local_dir, filename)` or NULL if file does not
  exist
  @keywords internal
}
\description{
Load from local drive
}
