% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dlw_call_list}
\alias{print.dlw_call_list}
\title{print gmd calls}
\usage{
\method{print}{dlw_call_list}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print gmd calls
}
