% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{query_or_form}
\alias{query_or_form}
\title{select between query or form according to method}
\usage{
query_or_form(method = c("GET", "POST"))
}
\arguments{
\item{method}{character: method of http request. Either "GET" or "POST".
Default is "GET".}
}
\value{
correspongin httr2 function
}
\description{
select between query or form according to method
}
\keyword{internal}
