% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo}
\alias{repo}
\alias{repo_create}
\alias{repo_load}
\title{dlw repository}
\usage{
repo_create(name, local_dir = getwd(), catalog = NULL)

repo_load(name, local_dir = getwd())
}
\arguments{
\item{name}{character: name of the repo (do NOT use spaces)}

\item{local_dir}{directory path}

\item{catalog}{data.table from [dlw_server_catalog]}
}
\value{
invisible data.table. It saves a dlw_repo file in `local_dir`

Loads dlw_repo file in `local_dir`
}
\description{
dlw repository
}
