% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{save_to_local}
\alias{save_to_local}
\title{save to local drive}
\usage{
save_to_local(x, local_dir, filename)
}
\arguments{
\item{x}{data in "dta" formar for now}

\item{local_dir}{local directory}

\item{filename}{filename. dta for now}
}
\value{
invisible(TRUE)
}
\description{
save to local drive
}
\keyword{internal}
