% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_server}
\alias{select_server}
\alias{select_collection}
\title{Select base server}
\usage{
select_server(server = NULL)

select_collection(server = NULL)
}
\arguments{
\item{server}{character: c("GMD"). Defaults to NULL (ie. GMD).}
}
\value{
character
}
\description{
Helper function to select server

Helper function to select collection argument for endpoint based on server
}
\keyword{internal}
