% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlwenv.R
\name{set_in_dlwenv}
\alias{set_in_dlwenv}
\title{Setter function: Assign a value in .dlwenv}
\usage{
set_in_dlwenv(key, value, verbose = getOption("dlw.verbose"))
}
\arguments{
\item{key}{A character string representing the key}

\item{value}{The value to store in .dlwenv}

\item{verbose}{logical: whether to display info}
}
\value{
The assigned value (invisibly)
}
\description{
Setter function: Assign a value in .dlwenv
}
\examples{
set_in_dlwenv("example_key", 42)
}
