% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simpleCap}
\alias{simpleCap}
\title{capitalize first letter in each word}
\usage{
simpleCap(s)
}
\arguments{
\item{s}{character}
}
\value{
first letter of each work in s capitalized
}
\description{
capitalize first letter in each word
}
\keyword{internal}
