% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_utils.R
\name{token}
\alias{token}
\alias{dlw_set_token}
\alias{dlw_get_token}
\alias{dlw_remove_token}
\alias{dlw_test_token}
\title{Datalibweb Token utilities}
\usage{
dlw_set_token(token)

dlw_get_token(prompt_if_missing = TRUE)

dlw_remove_token()

dlw_test_token(
  dlw_url = NULL,
  api_version = getOption("dlw.default_api_version"),
  verbose = getOption("dlw.verbose")
)
}
\arguments{
\item{token}{The API token as a string}

\item{prompt_if_missing}{logical: whether to be prompted to store token}

\item{dlw_url}{character: in case we have more than one url.}

\item{api_version}{character: API version}

\item{verbose}{logical: whether to display info}
}
\value{
invisible token

TRUE if all goes well
}
\description{
Datalibweb Token utilities

Store Datalibweb API token securely

Retrieve Datalibweb API token, prompt if missing/expired

Remove/reset the Datalibweb API token

confirm wither the token is working correctly
}
\examples{
\dontrun{
dlw_test_token()
}
}
\concept{token utilities}
