% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{build_request}
\alias{build_request}
\title{Build request version 2}
\usage{
build_request(
  dlw_url = NULL,
  server = NULL,
  api_version = getOption("dlw.default_api_version"),
  endpoint,
  ...
)
}
\arguments{
\item{dlw_url}{character: in case we have more than one url.}

\item{server}{character: in case we have more than one server. default GMD}

\item{api_version}{character: API version}

\item{endpoint}{character: dlw  API endpoint}

\item{...}{other parameters}
}
\value{
httr2 request
}
\description{
Build request version 2
}
