% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw-package.R
\docType{package}
\name{dlw-package}
\alias{dlw}
\alias{dlw-package}
\title{dlw: R Client for the internal Datalibweb API of the World Bank}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/worldbank/dlw}
  \item Report bugs at \url{https://github.com/worldbank/dlw/issues}
}

}
\author{
\strong{Maintainer}: R.Andres Castaneda \email{acastanedaa@worldbank.org}

}
\keyword{internal}
