% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_download}
\alias{dlw_download}
\title{Download data from datalibweb and save as a pin}
\usage{
dlw_download(
  country_code,
  filename,
  server = NULL,
  local_dir = getOption("dlw.local_dir"),
  local = fs::is_dir(local_dir),
  local_overwrite = FALSE,
  ...,
  verbose = getOption("dlw.verbose")
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{filename}{character: Name of the file to save/read (required)}

\item{server}{character: in case we have more than one server. default is  GMD}

\item{local_dir}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}

\item{local}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}

\item{local_overwrite}{logical. Whether to overwrite any saved data. Default
is FALSE}

\item{...}{additional filtering arguments (e.g.,survey_year, survey_acronym,
vermast, veralt, collection, module)}

\item{verbose}{logical: whether to display info}
}
\value{
Invisibly returns TRUE if download and pinning succeed
}
\description{
Download data from datalibweb and save as a pin
}
\keyword{internal}
