% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_download}
\alias{dlw_download}
\title{Download data from datalibweb and save as a pin}
\usage{
dlw_download(
  country_code,
  filename,
  board,
  pin_name,
  format,
  server = NULL,
  ...,
  verbose = getOption("dlw.verbose")
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{filename}{character: Name of the file to save/read (required)}

\item{board}{A pins board object (as returned by dlw_download)}

\item{pin_name}{The name of the pin (as returned by dlw_download)}

\item{format}{character: File format to use for pinning data ('qs'
[default] or 'parquet')}

\item{server}{character: in case we have more than one server. default is  GMD}

\item{...}{additional filtering arguments (e.g.,survey_year, survey_acronym,
vermast, veralt, collection, module)}

\item{verbose}{logical: whether to display info}

\item{board_type}{character: Which pins board to use:'local' (default
user-specific) 'folder' (shared network drive or similar).}
}
\value{
A list with the board and pin_name used
}
\description{
Download data from datalibweb and save as a pin
}
\keyword{internal}
