% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_data}
\alias{dlw_get_data}
\title{Get data from datalibweb (refactored)}
\usage{
dlw_get_data(
  country_code,
  filename,
  server = NULL,
  local_dir = getOption("dlw.local_dir"),
  local = fs::is_dir(local_dir),
  format = getOption("dlw.format"),
  local_overwrite = FALSE,
  version = NULL,
  verbose = getOption("dlw.verbose"),
  ...
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{filename}{character: Name of the file to save/read (required)}

\item{server}{character: in case we have more than one server. default is GMD}

\item{local_dir}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}

\item{local}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}

\item{format}{character: File format to use for saving data ('parquet'
[default] or 'qs2')}

\item{local_overwrite}{logical. Whether to overwrite any saved data. Default
is FALSE}

\item{version}{numeric: Version of the pin to read (for pinning data
retrieval only)}

\item{verbose}{logical: whether to display info}

\item{...}{additional filtering arguments (e.g.,survey_year, survey_acronym,
vermast, veralt, collection, module)}
}
\value{
data base request as data.table
}
\description{
`dlw_get_data()` function is the main user-facing function for 
retrieving datasets from the Datalibweb (DLW) API. It handles 
downloading, caching, and reading datasets, supporting both local 
and temporary storage.

**How it works:**
1. Checks if the requested file exists locally (unless `local_overwrite = TRUE`).
2. If it exists, reads it using `dlw_read()`.
3. If not, downloads the data from the DLW API using `dlw_download()`, saves 
   it in the specified format, and returns it as a `data.table`.
4. Handles directory management and caching via `get_wrk_board()`.

This function streamlines access to DLW datasets, automatically 
managing download, storage, and retrieval.
}
