% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_data}
\alias{dlw_get_data}
\title{get data from datalibweb}
\usage{
dlw_get_data(
  country_code,
  year = NULL,
  server = NULL,
  survey = NULL,
  module = NULL,
  verbose = getOption("dlw.verbose")
)
}
\arguments{
\item{survey}{}
}
\description{
get data from datalibweb
}
