% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_data}
\alias{dlw_get_data}
\title{get data from datalibweb (refactored)}
\usage{
dlw_get_data(
  country_code,
  server = NULL,
  filename,
  verbose = getOption("dlw.verbose"),
  local_dir = getOption("dlw.local_dir"),
  local = fs::is_dir(local_dir),
  board_type = c("folder", "local"),
  format = c("parquet", "qs"),
  local_overwrite = FALSE,
  ...
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{server}{character: in case we have more than one server. default is  GMD}

\item{filename}{character: Name of the file to save/read (required)}

\item{verbose}{logical: whether to display info}

\item{local_dir}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}

\item{local}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}

\item{board_type}{character: Which pins board to use: 'folder' (default,
shared) or 'local' (user-specific)}

\item{format}{character: File format to use for pinning data ('parquet'
[default] or 'qs')}

\item{local_overwrite}{logical. Whether to overwrite any saved data. Default
is FALSE}

\item{...}{additional filtering arguments (e.g.,survey_year, survey_acronym,
vermast, veralt, collection, module)}
}
\value{
data base request as data.table
}
\description{
get data from datalibweb (refactored)
}
