% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_data}
\alias{dlw_get_data}
\title{Get data from datalibweb (refactored)}
\usage{
dlw_get_data(
  country_code,
  filename,
  server = NULL,
  local_dir = getOption("dlw.local_dir"),
  local = fs::is_dir(local_dir),
  board_type = getOption("dlw.board_type"),
  format = getOption("dlw.format"),
  local_overwrite = FALSE,
  version = NULL,
  verbose = getOption("dlw.verbose"),
  ...
)
}
\arguments{
\item{country_code}{character: ISO3 code}

\item{filename}{character: Name of the file to save/read (required)}

\item{server}{character: in case we have more than one server. default is  GMD}

\item{local_dir}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}

\item{local}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}

\item{board_type}{character: Which pins board to use: 'folder' (default,
shared) or 'local' (user-specific). "local" Is useful if you want to share
pins between R sessions on your computer, and you don't care where the data
lives. In this case `local_dir` is ignored.}

\item{format}{character: File format to use for pinning data ('parquet'
[default] or 'qs')}

\item{local_overwrite}{logical. Whether to overwrite any saved data. Default
is FALSE}

\item{version}{numeric: Version of the pin to read (for pinning data
retrieval only)}

\item{verbose}{logical: whether to display info}

\item{...}{additional filtering arguments (e.g.,survey_year, survey_acronym,
vermast, veralt, collection, module)}
}
\value{
data base request as data.table
}
\description{
Get data from datalibweb (refactored)
}
