% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_data}
\alias{dlw_get_data}
\title{get data from datalibweb}
\usage{
dlw_get_data(server = NULL, country_code, year, collection, module)
}
\arguments{
\item{module}{}
}
\description{
get data from datalibweb
}
