% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_data}
\alias{dlw_get_data}
\title{get data from datalibweb}
\usage{
dlw_get_data(
  country_code,
  server = NULL,
  verbose = getOption("dlw.verbose"),
  filter = TRUE,
  local_dir = Sys.getenv("DLW_local_data"),
  local = fs::is_dir(local_dir),
  local_overwrite = FALSE,
  ...
)
}
\arguments{
\item{...}{additional filtering arguments (e.g.,survey_year, survey_acronym,
vermast, veralt, collection, module)}
}
\description{
get data from datalibweb
}
