% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_get_gmd}
\alias{dlw_get_gmd}
\title{Get GMD data.}
\usage{
dlw_get_gmd(
  country_code,
  year = NULL,
  module = NULL,
  survey = NULL,
  filename = NULL,
  vermast = NULL,
  veralt = NULL,
  latest = TRUE,
  ...
)
}
\arguments{
\item{country_code}{character: ISO3}

\item{year}{numeric: four digit year}

\item{module}{character: module of GMD collection (e.g., ALL, GPWG, L)}

\item{survey}{character: survey acronyn}

\item{filename}{character: File name}

\item{vermast}{character: Version of the master data in the form "vXX" where
X is a number of two digits like "01" or "02".}

\item{veralt}{character: Version of the alternative  data in the form "vXX"
where X is a number of two digits like "01" or "02".}

\item{latest}{logical. If TRUE and  `vermast` and `veralt` are NULL and then,
it will use the most recent data.}

\item{...}{
  Arguments passed on to \code{\link[=dlw_get_data]{dlw_get_data}}
  \describe{
    \item{\code{local_overwrite}}{}
  }}

\item{verbose}{logical. to display info.}
}
\value{
If the call is unique, it will return the data. If not, it will
  return the posibilities for the user to choose.
}
\description{
This is a wrapper of dlw_get_data. It just build the calls to dlw_get_data,
for ease of use.
}
\examples{
\dontrun{
  dlw_get_gmd("COL", 2010, "GPWG")
}
}
