% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_utils.R
\name{dlw_get_token}
\alias{dlw_get_token}
\title{Retrieve Datalibweb API token, prompt if missing/expired}
\usage{
dlw_get_token(prompt_if_missing = TRUE)
}
\arguments{
\item{prompt_if_missing}{logical: whether to be prompted to store token}
}
\value{
invisible token
}
\description{
Retrieve Datalibweb API token, prompt if missing/expired
}
