% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_read}
\alias{dlw_read}
\title{Read data from (local or temp)}
\usage{
dlw_read(dlw_dir, id_name, version = NULL)
}
\arguments{
\item{dlw_dir}{A folder object (as returned by dlw_download)}

\item{id_name}{The name of the a dataset (as returned by dlw_download)}

\item{version}{numeric: Version of the data to read (for versioning data
retrieval only)}
}
\value{
data.table
}
\description{
Read data from (local or temp)
}
\keyword{internal}
