% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_read}
\alias{dlw_read}
\title{Read data from a pin (local or temp)}
\usage{
dlw_read(board, pin_name, version = NULL)
}
\arguments{
\item{board}{A pins board object (as returned by dlw_download)}

\item{pin_name}{The name of the pin (as returned by dlw_download)}

\item{version}{numeric: Version of the pin to read (for pinning data
retrieval only)}
}
\value{
data.table
}
\description{
Read data from a pin (local or temp)
}
\keyword{internal}
