% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{dlw_read}
\alias{dlw_read}
\title{Read data from a pin (local or temp)}
\usage{
dlw_read(
  filename,
  local_dir = getOption("dlw.local_dir"),
  local = fs::is_dir(local_dir)
)
}
\arguments{
\item{filename}{character: Name of the file to save/read (required)}

\item{local_dir}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}

\item{local}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}
}
\value{
data.table
}
\description{
Read data from a pin (local or temp)
}
\keyword{internal}
