% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_server_catalog.R
\name{dlw_server_catalog}
\alias{dlw_server_catalog}
\title{Server catalog}
\usage{
dlw_server_catalog(
  server = NULL,
  dlw_url = NULL,
  api_version = getOption("dlw.default_api_version"),
  force = FALSE,
  verbose = getOption("dlw.verbose"),
  store_request = FALSE
)
}
\arguments{
\item{server}{character: in case we have more than one server. default is  GMD}

\item{dlw_url}{character: in case we have more than one url.}

\item{api_version}{character: API version}

\item{force}{logical: If TRUE, it will query the API regardless of whether
there is version available in .dlwenv}

\item{verbose}{logical: whether to display info}

\item{store_request}{logical: store request in .dlwenv as `last_req`}
}
\value{
dataframe with
}
\description{
Server catalog
}
\examples{
\dontrun{
dlw_server_catalog(sever = "GMD")
}
}
\seealso{
Other catalogs: 
\code{\link{dlw_country_catalog}()},
\code{\link{dlw_server_inventory}()}
}
\concept{catalogs}
