% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_server_catalog.R
\name{dlw_server_catalog}
\alias{dlw_server_catalog}
\title{Server catalog}
\usage{
dlw_server_catalog(
  server = NULL,
  dlw_url = NULL,
  api_version = getOption("dlw.default_api_version")
)
}
\arguments{
\item{server}{character: in case we have more than one server. default GMD}
}
\value{
dataframe with
}
\description{
Server catalog
}
\examples{
\dontrun{
dlw_server_catalog(sever = "GMD")
}
}
