% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlw_get_data.R
\name{get_wrk_board}
\alias{get_wrk_board}
\title{Get workign pips board}
\usage{
get_wrk_board(local, board_type, local_dir)
}
\arguments{
\item{local}{logical: whether or not to save and read data locally. default
is TRUE if `local_dir` exists.}

\item{board_type}{character: Which pins board to use: 'folder' (default,
shared) or 'local' (user-specific)}

\item{local_dir}{character: Local directory to save data. Default available
in option dlw.local_dir which is set initially as "".}
}
\value{
board from (pins) package
}
\description{
Get workign pips board
}
\keyword{internal}
