% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmd.R
\name{gmd_calls}
\alias{gmd_calls}
\title{get data calls when more than one option.}
\usage{
gmd_calls(ctl, country_code, ...)
}
\arguments{
\item{ctl}{data.table from [dlw_server_inventory]}

\item{country_code}{ISO3}
}
\value{
list of possible calls
}
\description{
get data calls when more than one option.
}
\seealso{
Other GMD utilities: 
\code{\link{dlw_get_gmd}()},
\code{\link{dlw_get_gmd_support}()},
\code{\link{eval_gmd_call}()}
}
\concept{GMD utilities}
\keyword{internal}
